move fake trajectory manager out
