+++ packages/dashboard/src/components/schedule-visualizer/tests/robot-trajectories-overlay.test.tsx
@@ -7,8 +7,8 @@ import React from 'react';
 import { robotHash } from 'react-components';
 import { act } from 'react-dom/test-utils';
 import { Map as LMap } from 'react-leaflet';
-import FakeTrajectoryManager from '../../../mock/fake-traj-manager';
-import { Conflict, Trajectory } from '../../../robot-trajectory-manager';
+import { Conflict, Trajectory } from '../../../managers/robot-trajectory-manager';
+import FakeTrajectoryManager from '../../../managers/__mocks__/robot-trajectory-manager';
 import { defaultSettings, TrajectoryAnimation } from '../../../settings';
 import { SettingsContext } from '../../app-contexts';
 import RobotTrajectoriesOverlay from '../robot-trajectories-overlay';
