+++ packages/dashboard/src/managers/__mocks__/trajectories.json
